--- BTCPayServer/Plugins/Multisig/Controllers/UIMultisigStatusController.cs
@@ -1,40 +1,0 @@
-#nullable enable
-
-using System;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using BTCPayServer.Abstractions.Constants;
-using BTCPayServer.Abstractions.Extensions;
-using BTCPayServer.Plugins.Multisig.Services;
-using BTCPayServer.Plugins.Wallets;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace BTCPayServer.Plugins.Multisig.Controllers;
-
-[Route("multisig-setups")]
-[Authorize(Policy = WalletPolicies.CanViewWallet, AuthenticationSchemes = AuthenticationSchemes.Cookie)]
-[Area(MultisigPlugin.Area)]
-public class UIMultisigStatusController(
-    MultisigService multisigService,
-    IAuthorizationService authorizationService) : Controller
-{
-    [HttpGet("{multisigSetupId}")]
-    public async Task<IActionResult> Status(string multisigSetupId)
-    {
-        var setupContext = await multisigService.GetPendingMultisigSetupContext(this.HttpContext.GetStoreData().Id, multisigSetupId);
-        var store = HttpContext.GetStoreDataOrNull();
-        if (setupContext is null || store is null)
-            return NotFound();
-
-        if (!setupContext.Pending.ReplacesExistingWallet && multisigService.HasOnChainWallet(store, setupContext.Pending.CryptoCode))
-            return NotFound();
-
-        var setupAccess = await authorizationService.GetSetupAccess(setupContext.Pending.StoreId, User, setupContext.Pending);
-        if (!setupAccess.CanViewStatus)
-            return Forbid();
-
-        var model = multisigService.CreateInProgressViewModel(setupContext.Pending.StoreId, User.GetId(), setupContext.Pending, setupAccess.CanManageWalletSettings);
-        return View("MultisigStatus", model);
-    }
-}
