+++ BTCPayServer/Plugins/Multisig/Views/SetupMultisigStatus.cshtml
@@ -90,11 +90,11 @@
     <div class="d-flex justify-content-end gap-3 mt-4">
         @if (Model.CanSubmitSignerKey)
         {
-            <a id="SubmitSignerKeyCta" asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigSignerKey" asp-action="SubmitMultisigSigner" asp-route-multisigSetupId="@Model.RequestId" class="btn btn-primary" text-translate="true">Submit signer key</a>
+            <a id="SubmitSignerKeyCta" asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigSetup" asp-action="SetupMultisigSubmitKey" asp-route-multisigSetupId="@Model.RequestId" class="btn btn-primary" text-translate="true">Submit signer key</a>
         }
         @if (Model.ReadyToCreateWallet && Model.CanCreateWallet)
         {
-            <form method="post" asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigSetup" asp-action="FinalizeMultisigSetup" asp-route-multisigSetupId="@Model.RequestId">
+            <form method="post" asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigWallets" asp-action="FinalizeMultisigSetup" asp-route-multisigSetupId="@Model.RequestId">
                 <button id="FinalizeMultisig" type="submit" class="btn btn-primary" text-translate="true">Create multisig</button>
             </form>
         }
