+++ BTCPayServer/Plugins/Multisig/Views/SetupWalletCard.cshtml
@@ -5,7 +5,7 @@
     <br>
     <div class="mt-5">
         <div class="list-group">
-            <a asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigSetup" asp-action="SetupMultisig" asp-route-storeId="@Model.StoreId" asp-route-cryptoCode="@Model.CryptoCode" id="ImportMultisigWizardLink" class="list-group-item list-group-item-action">
+            <a asp-area="@MultisigPlugin.Area" asp-controller="UIMultisigWallets" asp-action="SetupMultisig" asp-route-storeId="@Model.StoreId" asp-route-cryptoCode="@Model.CryptoCode" id="ImportMultisigWizardLink" class="list-group-item list-group-item-action">
                 <div class="image">
                     <vc:icon symbol="wallet-xpub"/>
                 </div>
